test: add coverage for match helper
